db_service/pyproject.toml
[tool.poetry]
name = "db-service"
version = "0.1.0"
description = ""
authors = ["legkovoyg <147488090+legkovoyg@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.38"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
